+++ src/main/java/futurefit2/core/interceptor/HttpLoggingInterceptor.java
@@ -31,10 +31,6 @@ public class HttpLoggingInterceptor implements Interceptor {
 
         Request request = chain.request();
 
-        request = request.newBuilder()//
-                .header("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:67.0) Gecko/20100101 Firefox/67.0")//
-                .build();
-
         HttpUrl url = request.url();
 
         Response proceed;