+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 0.3.0 - pending
 
 *	add custum logging implementation (based on retrofit1 implementation)
+*	add user agent interceptor with builder settings
 *	remove Futurefit 1 implementation
 
 ## 0.2.1 - 25/09/2019