+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 
 ## 0.3.0 - pending
 
-*	add custum logging implementation (based on retrofit1 implementation)
+*	add custom logging implementation (based on retrofit1 implementation)
 *	add user agent interceptor with builder settings
+*	add default okhttp3 cookieJar
 *	remove Futurefit 1 implementation
 
 ## 0.2.1 - 25/09/2019